Removal of Legacy Workbench Functions.
