+++ EpicLoot/Crafting/CraftSuccessDialog.cs
@@ -84,13 +84,64 @@ namespace EpicLoot.Crafting
         {
             gameObject.SetActive(false);
         }
+        public static ScrollRect ConvertToScrollingDescription(Text recipeDesc, Transform parent)
+        {
+            var contentSizeFitter = recipeDesc.gameObject.AddComponent<ContentSizeFitter>();
+            contentSizeFitter.verticalFit = ContentSizeFitter.FitMode.PreferredSize;
 
+            recipeDesc.resizeTextForBestFit = false;
+            recipeDesc.fontSize = 18;
+            recipeDesc.rectTransform.anchorMin = new Vector2(0, 1);
+            recipeDesc.rectTransform.anchorMax = new Vector2(1, 1); // pin top, stretch horiz
+            recipeDesc.rectTransform.pivot = new Vector2(0, 1);
+            recipeDesc.horizontalOverflow = HorizontalWrapMode.Wrap;
+            recipeDesc.rectTransform.anchoredPosition = new Vector2(4, 4);
+            recipeDesc.raycastTarget = false;
+
+            var scrollRectGO = new GameObject("ScrollView", typeof(RectTransform), typeof(ScrollRect), typeof(Image));
+            scrollRectGO.transform.SetParent(parent, false);
+            scrollRectGO.transform.SetSiblingIndex(0);
+            var rt = (RectTransform)scrollRectGO.transform;
+            rt.anchorMin = new Vector2(0, 1);
+            rt.anchorMax = new Vector2(0, 1);
+            rt.pivot = new Vector2(0, 1);
+            rt.anchoredPosition = new Vector2(11, -74);
+            rt.SetSizeWithCurrentAnchors(RectTransform.Axis.Horizontal, 330);
+            rt.SetSizeWithCurrentAnchors(RectTransform.Axis.Vertical, 300);
+            scrollRectGO.GetComponent<Image>().color = new Color(0, 0, 0, 0.2f);
+
+            var viewport = new GameObject("Viewport", typeof(RectTransform), typeof(RectMask2D));
+            viewport.transform.SetParent(scrollRectGO.transform, false);
+            var vrt = (RectTransform)viewport.transform;
+            vrt.anchorMin = new Vector2(0, 0);
+            vrt.anchorMax = new Vector2(1, 1);
+            vrt.sizeDelta = new Vector2(0, 0);
+            recipeDesc.transform.SetParent(vrt, false);
+
+            var scrollRect = scrollRectGO.GetComponent<ScrollRect>();
+            scrollRect.viewport = vrt;
+            scrollRect.content = recipeDesc.rectTransform;
+            scrollRect.horizontal = false;
+            scrollRect.vertical = true;
+            scrollRect.verticalScrollbarVisibility = ScrollRect.ScrollbarVisibility.AutoHide;
+            scrollRect.scrollSensitivity = 30;
+            scrollRect.inertia = false;
+            scrollRect.movementType = ScrollRect.MovementType.Clamped;
+            scrollRect.onValueChanged.RemoveAllListeners();
+
+            var newScrollbar = Object.Instantiate(InventoryGui.instance.m_recipeListScroll, scrollRectGO.transform);
+            newScrollbar.size = 0.4f;
+            scrollRect.onValueChanged.AddListener((_) => newScrollbar.size = 0.4f);
+            scrollRect.verticalScrollbar = newScrollbar;
+
+            return scrollRect;
+        }
         public static CraftSuccessDialog Create(Transform parent)
         {
             var inventoryGui = InventoryGui.instance;
             var newDialog = Instantiate(inventoryGui.m_variantDialog, parent);
             var dialog = newDialog.gameObject.AddComponent<CraftSuccessDialog>();
-            Destroy(newDialog);
+            //Destroy(newDialog);
             dialog.gameObject.name = "CraftingSuccessDialog";
 
             dialog.Frame = dialog.gameObject.transform.Find("VariantFrame").gameObject.RectTransform();
@@ -98,7 +149,7 @@ namespace EpicLoot.Crafting
             for (var i = 1; i < dialog.Frame.childCount; ++i)
             {
                 var child = dialog.Frame.GetChild(i);
-                Destroy(child.gameObject);
+                //Destroy(child.gameObject);
             }
             dialog.Frame.SetSizeWithCurrentAnchors(RectTransform.Axis.Horizontal, 380);
             dialog.Frame.SetSizeWithCurrentAnchors(RectTransform.Axis.Vertical, 550);
@@ -115,7 +166,7 @@ namespace EpicLoot.Crafting
             dialog.Description.rectTransform.SetSizeWithCurrentAnchors(RectTransform.Axis.Vertical, 460);
             dialog.Icon = Instantiate(inventoryGui.m_recipeIcon, dialog.Frame);
 
-            var scrollview = InventoryGui_UpdateRecipe_Patch.ConvertToScrollingDescription(dialog.Description, dialog.Frame);
+            var scrollview = ConvertToScrollingDescription(dialog.Description, dialog.Frame);
             var svrt = (RectTransform)scrollview.transform;
             svrt.SetSiblingIndex(1);
             svrt.anchorMin = new Vector2(0, 0);
@@ -127,7 +178,12 @@ namespace EpicLoot.Crafting
             var closeButton = dialog.gameObject.GetComponentInChildren<Button>();
             closeButton.onClick = new Button.ButtonClickedEvent();
             closeButton.onClick.AddListener(dialog.Close);
-            closeButton.transform.SetAsLastSibling();
+            Transform transform1;
+            (transform1 = closeButton.transform).SetAsLastSibling();
+            var cbrt = (RectTransform)transform1;
+            cbrt.anchoredPosition += new Vector2(0, -110);
+            
+            closeButton.gameObject.SetActive(true);
 
             return dialog;
         }
